The MainMenu block is placed in a separate component
